+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Remove temporal string on each JSX.escape, and make escape faster (#9c64ba4)
+- Skip escape for Int/Float attribute values (#8e8cf26)
+- Adds JSX.array (#3154012)
 
 ## 0.0.7